+++ test/scrypt-tests.js
@@ -1,5 +1,5 @@
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -  */
-/* Tests for scrypt key derivation function.                                   (c) C.Veness 2018  */
+/* Tests for scrypt key derivation function.                              (c) C.Veness 2018-2019  */
 /*                                                                                   MIT Licence  */
 /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -  */
 
@@ -12,43 +12,48 @@ const key0salt = 'c2NyeXB0AAwAAAAIAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
 
 
 describe('Scrypt tests', function() {
+    this.slow(20);
 
     describe('Hash & verify', function() {
+        this.slow(200); // kdf() is intentionally slow
+
         it('with just logN param', async function() {
-            const key = await Scrypt.kdf(password, { logN: 12 });
-            expect(Scrypt.viewParams(key)).to.deep.equal({ logN: 12, r: 8, p: 1 });
-            expect(await Scrypt.verify(key, password)).to.be.true;
+            const key = (await Scrypt.kdf(password, { logN: 12 })).toString('base64');
+            expect(Scrypt.viewParams(Buffer.from(key, 'base64'))).to.deep.equal({ logN: 12, r: 8, p: 1 });
+            expect(await Scrypt.verify(Buffer.from(key, 'base64'), password)).to.be.true;
         });
 
         it('with logN, r, p params', async function() {
-            const key = await Scrypt.kdf(password, { logN: 12, r: 9, p: 2 });
-            expect(Scrypt.viewParams(key)).to.deep.equal({ logN: 12, r: 9, p: 2 });
-            expect(await Scrypt.verify(key, password)).to.be.true;
+            const key = (await Scrypt.kdf(password, { logN: 12, r: 9, p: 2 })).toString('base64');
+            expect(Scrypt.viewParams(Buffer.from(key, 'base64'))).to.deep.equal({ logN: 12, r: 9, p: 2 });
+            expect(await Scrypt.verify(Buffer.from(key, 'base64'), password)).to.be.true;
         });
 
         it('with params as strings', async function() {
-            const key = await Scrypt.kdf(password, { logN: '12', r: '8', p: '1' });
-            expect(Scrypt.viewParams(key)).to.deep.equal({ logN: 12, r: 8, p: 1 });
-            expect(await Scrypt.verify(key, password)).to.be.true;
+            const key = (await Scrypt.kdf(password, { logN: '12', r: '8', p: '1' })).toString('base64');
+            expect(Scrypt.viewParams(Buffer.from(key, 'base64'))).to.deep.equal({ logN: 12, r: 8, p: 1 });
+            expect(await Scrypt.verify(Buffer.from(key, 'base64'), password)).to.be.true;
         });
 
         it('fails to verify with bad passphrase', async function() {
-            const key = await Scrypt.kdf(password, { logN: 12 });
-            expect(await Scrypt.verify(key, 'wrong password')).to.be.false;
+            const key = (await Scrypt.kdf(password, { logN: 12 })).toString('base64');
+            expect(await Scrypt.verify(Buffer.from(key, 'base64'), 'wrong password')).to.be.false;
         });
     });
 
     describe('Verify previous key', function() {
         it('verifies null-salt key', async function() {
-            expect(await Scrypt.verify(key0salt, password)).to.be.true;
+            expect(await Scrypt.verify(Buffer.from(key0salt, 'base64'), password)).to.be.true;
         });
 
         it('fails to verify null-salt key with bad passphrase', async function() {
-            expect(await Scrypt.verify(key0salt, 'wrong password')).to.be.false;
+            expect(await Scrypt.verify(Buffer.from(key0salt, 'base64'), 'wrong password')).to.be.false;
         });
     });
 
     describe('TypedArray/Buffer passphrase', function() {
+        this.slow(200); // kdf() is intentionally slow
+
         it('Uint8Array', async function() {
             const pwTypedArray = new Uint8Array([ 99, 98, 97, 96, 95, 94, 94, 92, 91 ]);
             const key = await Scrypt.kdf(pwTypedArray, { logN: 12 });
@@ -66,7 +71,7 @@ describe('Scrypt tests', function() {
 
     describe('Pick params', function() {
         it('Picks params for 100ms', async function() {
-            const params = await Scrypt.pickParams(0.1, 1024*1024*1024, 0.5);
+            const params = Scrypt.pickParams(0.1, 1024*1024*1024, 0.5);
             expect(params).to.have.all.keys('logN', 'r', 'p');
             expect(params.logN).to.be.within(8, 20);
             expect(params.r).to.equal(8);
@@ -74,7 +79,7 @@ describe('Scrypt tests', function() {
         });
 
         it('Picks params with default maxmem/maxmemfrac', async function() {
-            const params = await Scrypt.pickParams(0.1);
+            const params = Scrypt.pickParams(0.1);
             expect(params).to.have.all.keys('logN', 'r', 'p');
             expect(params.logN).to.be.within(8, 20);
             expect(params.r).to.equal(8);
@@ -82,17 +87,17 @@ describe('Scrypt tests', function() {
         });
 
         it('Picks params with 0 maxmem', async function() {
-            const params = await Scrypt.pickParams(0.1, 0);
+            const params = Scrypt.pickParams(0.1, 0);
             expect(params.logN).to.be.within(8, 20);
         });
 
         it('Picks params with 0 maxmemfrac', async function() {
-            const params = await Scrypt.pickParams(0.1, 0, 0);
+            const params = Scrypt.pickParams(0.1, 0, 0);
             expect(params.logN).to.be.within(8, 20);
         });
 
         it('Picks params setting N based on memory limit', async function() {
-            const params = await Scrypt.pickParams(1, 1024, 0.1);
+            const params = Scrypt.pickParams(1, 1024, 0.1);
             expect(params.logN).to.be.within(8, 20);
             expect(params.p).to.be.above(1);
         });
@@ -157,23 +162,22 @@ describe('Scrypt tests', function() {
                 .catch(error => expect(error.message).to.equal('Passphrase must be a string, TypedArray, or Buffer')));
             it('throws on bad key type', () => Scrypt.verify(null, 'passwd')
                 .then(() => { throw new Error('Test should fail'); })
-                .catch(error => expect(error.message).to.equal('Key must be a string')));
-            it('throws on bad key', () => Scrypt.verify('key', 'passwd')
+                .catch(error => expect(error.message).to.equal('Key must be a Buffer')));
+            it('throws on bad key', () => Scrypt.verify(Buffer.from('key', 'base64'), 'passwd')
                 .then(() => { throw new Error('Test should fail'); })
                 .catch(error => expect(error.message).to.equal('Invalid key')));
             it('fails to verify on checksum failure', async () => {
-                const keyBuff = Buffer.from(await Scrypt.kdf(password, { logN: 12 }), 'base64');
-                keyBuff[7] = 11; // patch logN to new value
-                const keybis = keyBuff.toString('base64');
-                expect(Scrypt.viewParams(keybis)).to.deep.equal({ logN: 11, r: 8, p: 1 });
-                expect(await Scrypt.verify(keybis, password)).to.be.false;
+                const key = await Scrypt.kdf(password, { logN: 12 });
+                key[7] = 11; // patch logN to new value
+                expect(Scrypt.viewParams(key)).to.deep.equal({ logN: 11, r: 8, p: 1 });
+                expect(await Scrypt.verify(key, password)).to.be.false;
             });
         });
 
         describe('viewParams errors', function() { // note Scrypt.viewParams is not async
-            it('throws on null key', () => expect(() => Scrypt.viewParams(null)).to.throw(TypeError, 'Key must be a string'));
-            it('throws on numeric key', () => expect(() => Scrypt.viewParams(99)).to.throw(TypeError, 'Key must be a string'));
-            it('throws on invalid key', () => expect(() => Scrypt.viewParams('bad key')).to.throw(RangeError, 'Invalid key'));
+            it('throws on null key', () => expect(() => Scrypt.viewParams(null)).to.throw(TypeError, 'Key must be a Buffer'));
+            it('throws on numeric key', () => expect(() => Scrypt.viewParams(99)).to.throw(TypeError, 'Key must be a Buffer'));
+            it('throws on invalid key', () => expect(() => Scrypt.viewParams(Buffer.from('bad key', 'base64'))).to.throw(RangeError, 'Invalid key'));
         });
 
     });